セッションと認証関連の実装

+++ src/gin_sample/service/book.go
@@ -20,7 +20,7 @@ type BookService struct {
 
 //Book bookテーブル
 type Book struct {
-	ID      int64  `xorm:"'id'"`
+	ID      int    `xorm:"'id'"`
 	title   string `xorm:"'title'"`
 	content string `xorm:"'content'"`
 }
